changesssss

+++ commandhandler.py
@@ -179,6 +179,8 @@ class commandhandler:
 			else:
 				if args[1].isnumeric:
 					for i in range(0,int(args[1])):
+						if len(self.last_MSG == 0): # if arg is higher than number of stored messages that can be deleted
+							break
 						await self.last_MSG.pop().delete()
 				else:
 					await self.last_MSG.pop().delete()
